y now int

+++ 01-10/01-10.cpp
@@ -22,7 +22,7 @@ int main()
 
 	int x;
 	cin >> x;
-	double y;	
+	int y;	
 	cin >> y;
 
 	double xd, yd;
